LayoutTests/imported/w3c/web-platform-tests/html/canvas/element/pixel-manipulation/2d.imageData.create2.round-expected.txt

2d.imageData.create2.round
createImageData(w, h) is rounded the same as getImageData(0, 0, w, h)
Actual output:

PASS createImageData(w, h) is rounded the same as getImageData(0, 0, w, h)
